+++ BlockBreaker2024/core/src/main/java/puppy/code/Paddle.java
@@ -11,7 +11,7 @@ public class Paddle extends GameObject {
     private int width = 100;
     private int height = 10;
     */
-    public Paddle(int x, int y, int width, int height) {
+    public Paddle(float x, float y, float width, float height) {
     	super(x, y, width, height);
     }
     
@@ -24,7 +24,7 @@ public class Paddle extends GameObject {
 
 	public void draw(ShapeRenderer shape){
 		shape.setColor(Color.BLUE);
-        int newX = x;
+		float newX = x;
 
         if (Gdx.input.isKeyPressed(Input.Keys.LEFT)) newX = x - 15;
         if (Gdx.input.isKeyPressed(Input.Keys.RIGHT)) newX = x + 15;
